+++ src/packages/core/validation/context/index.ts
@@ -1,2 +1,4 @@
 export * from './validation.context.js';
 export * from './validation.context-token.js';
+export * from './server-model-validator.context-token.js';
+export * from './server-model-validator.context.js';
