+++ src/packages/core/validation/context/server-model-validator.context.ts
@@ -1,6 +1,6 @@
 import type { UmbValidator } from '../interfaces/validator.interface.js';
 import { UmbDataPathPropertyValueQuery } from '../utils/index.js';
-import { UMB_VALIDATION_CONTEXT } from '../context/validation.context-token.js';
+import { UMB_VALIDATION_CONTEXT } from './validation.context-token.js';
 import { UMB_VALIDATION_EMPTY_LOCALIZATION_KEY } from '../const.js';
 import { UMB_SERVER_MODEL_VALIDATOR_CONTEXT } from './server-model-validator.context-token.js';
 import { UmbContextBase } from '@umbraco-cms/backoffice/class-api';
