+++ src/packages/core/validation/context/validation.context-token.ts
@@ -1,4 +1,4 @@
-import type { UmbValidationContext } from './validation.context.js';
+import type { UmbValidationController } from '../controllers/validation.controller.js';
 import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
 
-export const UMB_VALIDATION_CONTEXT = new UmbContextToken<UmbValidationContext>('UmbValidationContext');
+export const UMB_VALIDATION_CONTEXT = new UmbContextToken<UmbValidationController>('UmbValidationContext');
