+++ src/packages/core/validation/controllers/index.ts
@@ -1,5 +1,4 @@
 export * from './bind-server-validation-to-form-control.controller.js';
 export * from './form-control-validator.controller.js';
 export * from './observe-validation-state.controller.js';
-export * from './server-model-validator.context-token.js';
-export * from './server-model-validator.context.js';
+export * from './validation.controller.js';
